SOLR-14654 Remove plugin loading from .system collection (for 9.0) (#1677)
